COMP: Signedness mismatch
This changes the cases when an expression of the signed type
A is used in the context of unsigned type B (assigned, casted, used in a
parameter, returned from a method, function or a block, etc.).

+++ src/metaArrow.cxx
@@ -208,7 +208,7 @@ MetaArrow::M_SetupWriteFields()
   m_Fields.push_back(mF);
 
   mF = new MET_FieldRecordType;
-  MET_InitWriteField(mF, "Direction", MET_DOUBLE_ARRAY, m_NDims, M_Direction);
+  MET_InitWriteField(mF, "Direction", MET_DOUBLE_ARRAY, static_cast<size_t>(m_NDims), M_Direction);
   m_Fields.push_back(mF);
 }
