+++ src/metaForm.h
@@ -142,11 +142,11 @@ public:
                int               _dependsOn = -1)
   {
     MET_FieldRecordType * mFw = new MET_FieldRecordType;
-    MET_InitWriteField(mFw, _fieldName, _type, _length, _v);
+    MET_InitWriteField(mFw, _fieldName, _type, static_cast<size_t>(_length), _v);
     m_UserDefinedWriteFields.push_back(mFw);
 
     MET_FieldRecordType * mFr = new MET_FieldRecordType;
-    MET_InitReadField(mFr, _fieldName, _type, _required, _dependsOn, _length);
+    MET_InitReadField(mFr, _fieldName, _type, _required, _dependsOn, static_cast<size_t>(_length));
     m_UserDefinedReadFields.push_back(mFr);
 
     return true;
